+++ set-objects-writable/find-readonly-objects.sh
@@ -8,7 +8,7 @@
 # Copyright 2017 Fwolf <fwolf.aide+git-hooks@gmail.com>
 # Distributed under the MIT license.
 #
-# Last Modified: 2017-08-17T16:31:19+08:00, r19
+# Last Modified: 2017-08-17T16:50:33+08:00, r20
 #====================================================================