+++ set-objects-writable/install.sh
@@ -3,7 +3,7 @@
 # Copyright 2017 Fwolf <fwolf.aide+git-hooks@gmail.com>
 # Distributed under the MIT license.
 #
-# Last Modified: 2017-08-17T12:44:49+08:00, r17
+# Last Modified: 2017-08-17T16:50:33+08:00, r20
 #====================================================================
 
 
@@ -20,25 +20,21 @@ findGitRepositoryDir "$GIT_ROOT"
 
 cp -r "$SCRIPT_DIR"../inc/ "$GIT_REPO_DIR"hooks/
 cp "$SCRIPT_DIR"../fix-permissions.sh "$GIT_REPO_DIR"hooks/
-cp "$SCRIPT_DIR"updateLastModifiedTime.php "$GIT_REPO_DIR"hooks/
-cp "$SCRIPT_DIR"autoLastModifiedTime.sh "$GIT_REPO_DIR"hooks/
+cp "$SCRIPT_DIR"find-readonly-objects.sh "$GIT_REPO_DIR"hooks/
+cp "$SCRIPT_DIR"set-objects-writable.sh "$GIT_REPO_DIR"hooks/
 
 PWD_BAK="$PWD"
 cd "$GIT_REPO_DIR"hooks/
 chmod +x fix-permissions.sh \
-    updateLastModifiedTime.php \
-    autoLastModifiedTime.sh
+    find-readonly-objects.sh \
+    set-objects-writable.sh
 chmod +x inc/*
 
 
-# Special call with 'source' and exit code check when commit maybe rejected
-HOOK_CONTENT=". \${0%/*}/autoLastModifiedTime.sh
-if [ 0 -ne \${EXIT_CODE} ]; then
-    echo Commit rejected by autoLastModifiedTime hook
-    exit 1
-fi"
+HOOK_CONTENT="\${0%/*}/set-objects-writable.sh"
 
-installHook pre-commit "$HOOK_CONTENT" 1l0pmscuzl1szqu7
+installHook post-commit $HOOK_CONTENT 1l0ebpnpj11s5j5p
+installHook post-merge $HOOK_CONTENT 1l0ebpnpj11s5j5p
 
 
 cd "$PWD_BAK"