+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 
 
+### Un-versioned
+
+- Change filename from camelCase to with dashes
+
+
 ### v0.1 (2017-08-17)
 
 - New script fix-permission.sh to help set hook and script permission